requirements.txt
alembic==1.14.0
anyio==4.7.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
autopep8==2.3.1
azure-core==1.32.0
azure-cosmos==4.3.1
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
babel==2.16.0
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.2.0
blinker==1.9.0
boto3==1.35.87
botocore==1.35.87
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
databricks-connect==15.4.4
databricks-feature-engineering==0.6.0
databricks-feature-lookup==1.2.0
databricks-sdk==0.32.0
dbl-tempo==0.1.29
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.15
dill==0.3.9
docker==7.1.0
executing==2.1.0
fastjsonschema==2.21.1
flask==2.3.3
fonttools==4.55.3
fqdn==1.5.1
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
grpcio==1.68.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.8
lightgbm==4.5.0
mako==1.3.8
markdown==3.7
markupsafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mlflow==2.17.0
mlflow-skinny==2.17.0
mypy==1.14.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nbqa==1.9.1
nest-asyncio==1.6.0
notebook==7.3.2
notebook-shim==0.2.4
numpy==1.26.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
pip==24.3.1
pip-tools==7.4.1
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.7
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pygments==2.18.0
pylint==3.3.3
pymysql==1.1.1
pyparsing==3.2.0
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
send2trash==1.8.3
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenize-rt==6.1.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
zipp==3.21.0
